+++ rockpool/typehints.py
@@ -8,6 +8,7 @@ See :ref:`/reference/params_types.ipynb` for more information on the available t
 import numpy as np
 from typing import Union, Any, Callable, Dict
 from collections import abc
+from jax import numpy as jnp
 
 from rockpool.parameters import ParameterBase
 
@@ -61,3 +62,11 @@ JaxRNGKey = Any
 JaxTreeDef = Any
 
 TreeDef = Dict
+
+
+class DRCError(ValueError):
+    pass
+
+
+class DRCWarning(Warning, DRCError):
+    pass